refactor and better message handling
